Add step parsing functionality and tests
- Implement step parsing logic to handle various step types.
- Refactor mapping functions to utilize the new parser.
- Add unit tests for step parsing (without framework for the moment)

+++ running_addict_plan_extractor/model/model.py
@@ -10,7 +10,7 @@ class BaseStep:
 
 @dataclass
 class ConstantStep(BaseStep):
-    duration_minutes: int
+    duration_minutes: float
     pace: Pace
 
 
@@ -23,7 +23,7 @@ class IntervalStep(BaseStep):
 
 @dataclass
 class ProgressiveStep(BaseStep):
-    duration_minutes: int
+    duration_minutes: float
     start_pace: Pace
     end_pace: Pace
